# CodeGen/aspnet/src/Org.OpenAPITools/wwwroot/openapi-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Vaughan Community Church Social Network",
    "license" : {
      "name" : "MIT"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://api.vccc.ca/v1"
  } ],
  "paths" : {
    "/users" : {
      "get" : {
        "tags" : [ "users" ],
        "summary" : "List all users",
        "operationId" : "listUsers",
        "parameters" : [ {
          "name" : "limit",
          "in" : "query",
          "description" : "How many items to return at one time (max 100)",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "A paged array of users",
            "headers" : {
              "x-next" : {
                "description" : "A link to the next page of responses",
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Users"
                }
              }
            }
          },
          "default" : {
            "description" : "unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "users" ],
        "summary" : "Create a user",
        "operationId" : "createUsers",
        "responses" : {
          "201" : {
            "description" : "Null response"
          },
          "default" : {
            "description" : "unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/users/{userId}" : {
      "get" : {
        "tags" : [ "usres" ],
        "summary" : "Info for a specific user",
        "operationId" : "showUserById",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "The id of the user to retrieve",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Expected response to a valid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Users"
                }
              }
            }
          },
          "default" : {
            "description" : "unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "User" : {
        "required" : [ "id", "name" ],
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "tag" : {
            "type" : "string"
          }
        },
        "example" : {
          "name" : "name",
          "id" : 0,
          "tag" : "tag"
        }
      },
      "Users" : {
        "type" : "array",
        "items" : {
          "$ref" : "#/components/schemas/User"
        }
      },
      "Error" : {
        "required" : [ "code", "message" ],
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          }
        }
      }
    }
  }
}